Create 20200520040427-AddInsertPlanterOrgIDTrigger.js
Created a trigger that will update planter_org in tree table to match the planter_org_id of planter table. This migration was tested - however limited unit testing was preformed due to ablitiy to create triggers in treestand database and test database was not an exact match to treetracker database.